Replace int special1/special2 with a union
The way it was implemented before looks like a fucking kludge: int
special1/special2 variables could store either an integer value or a
pointer to struct mobj_s or struct player_s. This could work on a 32-
bit architectures where the int and pointer size was 4 bytes, but on
fucking 64-bit architectures the pointer size is 8 bytes and int is still 4 bytes.